+++ cassandra/cqlengine/connection.py
@@ -98,7 +98,13 @@ class Connection(object):
         if self.lazy_connect:
             return
 
-        self.cluster = Cluster(self.hosts, **self.cluster_options)
+        if 'cloud' in self.cluster_options:
+            if self.hosts:
+                log.warning("Ignoring hosts %s because a cloud config was provided.", self.hosts)
+            self.cluster = Cluster(**self.cluster_options)
+        else:
+            self.cluster = Cluster(self.hosts, **self.cluster_options)
+
         try:
             self.session = self.cluster.connect()
             log.debug(format_log_context("connection initialized with internally created session", connection=self.name))
@@ -301,6 +307,8 @@ def set_session(s):
     log.debug("cqlengine default connection initialized with %s", s)
 
 
+# TODO next major: if a cloud config is specified in kwargs, hosts will be ignored.
+# This function should be refactored to reflect this change. PYTHON-1265
 def setup(
         hosts,
         default_keyspace,